Landing Page, base.php, further styling
Landing Page is complete. I made a base.php file which will serve as the
template for any new pages we decide to create. For example, when we
make the doctor page we will just make a copy of base.php, rename it,
and start coding inside of the new doctor.php file or whatever we decide
to call it. I will simplify base a little more to utilize includes to
don't use it quite yet. Please let me know about any UI issues you
notice.