base.php simplified
base.php is used to make any new pages (i.e., it keeps the template
consistent across all pages). If you're going to make a new page, make a
copy of this file and rename it. In the templates folder, the header,
footer, and meta php files are included in the base.php file. Changes
made to these files will reflect across all pages.